requirements.txt
pillow>=11.2.1
reportlab>=4.1.0
markdown>=3.5
beautifulsoup4>=4.12.2
pygments>=2.18
markdownify>=0.12
